requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
openai==1.58.1
qdrant-client==1.12.1
pypdf==5.1.0
pydantic==2.10.4
httpx==0.28.1
